feat: Small refactor addressing comments [DHIS2-20256]

+++ dhis-2/dhis-services/dhis-service-reporting/src/main/java/org/hisp/dhis/document/impl/DefaultDocumentService.java
@@ -35,7 +35,11 @@ import lombok.RequiredArgsConstructor;
 import org.hisp.dhis.document.Document;
 import org.hisp.dhis.document.DocumentService;
 import org.hisp.dhis.document.DocumentStore;
+import org.hisp.dhis.feedback.ForbiddenException;
+import org.hisp.dhis.fileresource.FileResource;
 import org.hisp.dhis.fileresource.FileResourceService;
+import org.hisp.dhis.security.acl.AclService;
+import org.hisp.dhis.user.CurrentUserUtil;
 import org.hisp.dhis.user.User;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -48,6 +52,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Service("org.hisp.dhis.document.DocumentService")
 public class DefaultDocumentService implements DocumentService {
   private final FileResourceService fileResourceService;
+  private final AclService aclService;
 
   private final DocumentStore documentStore;
 
@@ -73,8 +78,19 @@ public class DefaultDocumentService implements DocumentService {
   }
 
   @Override
-  public void deleteDocument(Document document) {
+  @Transactional
+  public void deleteDocument(Document document) throws ForbiddenException {
+    if (!aclService.canDelete(CurrentUserUtil.getCurrentUserDetails(), document)) {
+      throw new ForbiddenException("You don't have the proper permissions to delete this object.");
+    }
+
+    // unassign the doc's file resource when deleting
+    FileResource fr = document.getFileResource();
     documentStore.delete(document);
+    if (fr != null) {
+      fr.setAssigned(false);
+      fileResourceService.updateFileResource(fr);
+    }
   }
 
   @Override
